+++ src/main/java/com/balybus/galaxy/chat/service/ChatServiceImpl.java
@@ -17,8 +17,8 @@ import com.balybus.galaxy.careAssistant.repository.HelperRepository;
 import com.balybus.galaxy.login.classic.domain.type.RoleType;
 import com.balybus.galaxy.member.domain.TblUser;
 import com.balybus.galaxy.member.repository.MemberRepository;
-import com.balybus.galaxy.patient.domain.tblPatientLog.TblPatientLog;
-import com.balybus.galaxy.patient.domain.tblPatientLog.TblPatientLogRepository;
+import com.balybus.galaxy.global.domain.tblPatientLog.TblPatientLog;
+import com.balybus.galaxy.global.domain.tblPatientLog.TblPatientLogRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
