+++ src/main/java/com/balybus/galaxy/global/domain/tblChatRoom/TblChatRoom.java
@@ -2,7 +2,7 @@ package com.balybus.galaxy.global.domain.tblChatRoom;
 
 import com.balybus.galaxy.global.domain.BaseEntity;
 import com.balybus.galaxy.member.domain.TblUser;
-import com.balybus.galaxy.patient.domain.tblPatientLog.TblPatientLog;
+import com.balybus.galaxy.global.domain.tblPatientLog.TblPatientLog;
 import jakarta.persistence.*;
 import lombok.*;
 import org.hibernate.annotations.Comment;
