+++ src/main/java/com/balybus/galaxy/global/domain/tblMatching/MatchingServiceImpl.java
@@ -9,8 +9,8 @@ import com.balybus.galaxy.global.utils.mail.dto.contents.ContentDto;
 import com.balybus.galaxy.global.utils.mail.dto.contents.MailMatchingDto;
 import com.balybus.galaxy.careAssistant.domain.TblHelper;
 import com.balybus.galaxy.careAssistant.repository.HelperRepository;
-import com.balybus.galaxy.patient.domain.tblPatientLog.TblPatientLog;
-import com.balybus.galaxy.patient.domain.tblPatientLog.TblPatientLogRepository;
+import com.balybus.galaxy.global.domain.tblPatientLog.TblPatientLog;
+import com.balybus.galaxy.global.domain.tblPatientLog.TblPatientLogRepository;
 import jakarta.mail.MessagingException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
