+++ src/main/java/com/balybus/galaxy/global/domain/tblPatientLog/TblPatientLogRepository.java
@@ -1,11 +1,10 @@
-package com.balybus.galaxy.patient.domain.tblPatientLog;
+package com.balybus.galaxy.global.domain.tblPatientLog;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
-import java.util.Optional;
 
 public interface TblPatientLogRepository extends JpaRepository<TblPatientLog, Long> {
     Page<TblPatientLog> findByManagerId(Long managerSeq, Pageable page);
