+++ src/main/java/com/balybus/galaxy/patient/PatientBaseDto.java
@@ -1,10 +1,10 @@
-package com.balybus.galaxy.patient.dto.baseDto;
+package com.balybus.galaxy.patient;
 
 import com.balybus.galaxy.global.domain.tblCare.dto.CareBaseDto;
-import com.balybus.galaxy.patient.domain.tblPatient.TblPatient;
-import com.balybus.galaxy.patient.domain.tblPatientLog.TblPatientLog;
-import com.balybus.galaxy.patient.domain.tblPatientTime.TblPatientTime;
-import com.balybus.galaxy.patient.domain.tblPatientTimeLog.TblPatientTimeLog;
+import com.balybus.galaxy.global.domain.tblPatient.TblPatient;
+import com.balybus.galaxy.global.domain.tblPatientLog.TblPatientLog;
+import com.balybus.galaxy.global.domain.tblPatientTime.TblPatientTime;
+import com.balybus.galaxy.global.domain.tblPatientTimeLog.TblPatientTimeLog;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
