+++ src/main/java/com/balybus/galaxy/patient/basic/dto/BasicRequestDto.java
@@ -4,8 +4,8 @@ import com.balybus.galaxy.global.domain.tblAddressFirst.TblAddressFirst;
 import com.balybus.galaxy.global.domain.tblAddressSecond.TblAddressSecond;
 import com.balybus.galaxy.global.domain.tblAddressThird.TblAddressThird;
 import com.balybus.galaxy.global.domain.tblCenterManager.TblCenterManager;
-import com.balybus.galaxy.patient.domain.tblPatient.TblPatient;
-import com.balybus.galaxy.patient.dto.baseDto.PatientBaseDto;
+import com.balybus.galaxy.global.domain.tblPatient.TblPatient;
+import com.balybus.galaxy.patient.PatientBaseDto;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
