+++ src/main/java/com/balybus/galaxy/patient/basic/dto/BasicResponseDto.java
@@ -1,9 +1,9 @@
 package com.balybus.galaxy.patient.basic.dto;
 
 import com.balybus.galaxy.global.domain.tblCare.dto.CareChoiceListDto;
-import com.balybus.galaxy.patient.domain.tblPatient.TblPatient;
-import com.balybus.galaxy.patient.domain.tblPatientTime.TblPatientTime;
-import com.balybus.galaxy.patient.dto.baseDto.PatientBaseDto;
+import com.balybus.galaxy.global.domain.tblPatient.TblPatient;
+import com.balybus.galaxy.global.domain.tblPatientTime.TblPatientTime;
+import com.balybus.galaxy.patient.PatientBaseDto;
 import lombok.Builder;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
