+++ src/main/java/com/balybus/galaxy/patient/basic/service/BasicServiceImpl.java
@@ -15,11 +15,11 @@ import com.balybus.galaxy.global.utils.file.service.FileService;
 import com.balybus.galaxy.login.classic.service.loginAuth.LoginAuthCheckServiceImpl;
 import com.balybus.galaxy.patient.basic.dto.BasicRequestDto;
 import com.balybus.galaxy.patient.basic.dto.BasicResponseDto;
-import com.balybus.galaxy.patient.domain.tblPatient.TblPatient;
-import com.balybus.galaxy.patient.domain.tblPatient.TblPatientRepository;
-import com.balybus.galaxy.patient.domain.tblPatientTime.TblPatientTime;
-import com.balybus.galaxy.patient.domain.tblPatientTime.TblPatientTimeRepository;
-import com.balybus.galaxy.patient.dto.baseDto.PatientBaseDto;
+import com.balybus.galaxy.global.domain.tblPatient.TblPatient;
+import com.balybus.galaxy.global.domain.tblPatient.TblPatientRepository;
+import com.balybus.galaxy.global.domain.tblPatientTime.TblPatientTime;
+import com.balybus.galaxy.global.domain.tblPatientTime.TblPatientTimeRepository;
+import com.balybus.galaxy.patient.PatientBaseDto;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
