+++ src/main/java/com/balybus/galaxy/patient/matchingStatus/service/MatchingStatusServiceImpl.java
@@ -8,8 +8,8 @@ import com.balybus.galaxy.global.domain.tblMatching.TblMatching;
 import com.balybus.galaxy.global.domain.tblMatching.TblMatchingRepository;
 import com.balybus.galaxy.global.exception.BadRequestException;
 import com.balybus.galaxy.login.classic.service.loginAuth.LoginAuthCheckServiceImpl;
-import com.balybus.galaxy.patient.domain.tblPatientLog.TblPatientLog;
-import com.balybus.galaxy.patient.domain.tblPatientLog.TblPatientLogRepository;
+import com.balybus.galaxy.global.domain.tblPatientLog.TblPatientLog;
+import com.balybus.galaxy.global.domain.tblPatientLog.TblPatientLogRepository;
 import com.balybus.galaxy.patient.matchingStatus.dto.MatchingStatusResponseDto;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
