+++ src/main/java/com/balybus/galaxy/patient/recruit/service/RecruitServiceImpl.java
@@ -16,14 +16,14 @@ import com.balybus.galaxy.global.utils.address.service.serviceImpl.TblAddressThi
 import com.balybus.galaxy.global.utils.code.CodeServiceImpl;
 import com.balybus.galaxy.global.utils.file.service.FileService;
 import com.balybus.galaxy.login.classic.service.loginAuth.LoginAuthCheckServiceImpl;
-import com.balybus.galaxy.patient.domain.tblPatient.TblPatient;
-import com.balybus.galaxy.patient.domain.tblPatientLog.TblPatientLog;
-import com.balybus.galaxy.patient.domain.tblPatientLog.TblPatientLogRepository;
-import com.balybus.galaxy.patient.domain.tblPatientTime.TblPatientTime;
-import com.balybus.galaxy.patient.domain.tblPatientTime.TblPatientTimeRepository;
-import com.balybus.galaxy.patient.domain.tblPatientTimeLog.TblPatientTimeLog;
-import com.balybus.galaxy.patient.domain.tblPatientTimeLog.TblPatientTimeLogRepository;
-import com.balybus.galaxy.patient.dto.baseDto.PatientBaseDto;
+import com.balybus.galaxy.global.domain.tblPatient.TblPatient;
+import com.balybus.galaxy.global.domain.tblPatientLog.TblPatientLog;
+import com.balybus.galaxy.global.domain.tblPatientLog.TblPatientLogRepository;
+import com.balybus.galaxy.global.domain.tblPatientTime.TblPatientTime;
+import com.balybus.galaxy.global.domain.tblPatientTime.TblPatientTimeRepository;
+import com.balybus.galaxy.global.domain.tblPatientTimeLog.TblPatientTimeLog;
+import com.balybus.galaxy.global.domain.tblPatientTimeLog.TblPatientTimeLogRepository;
+import com.balybus.galaxy.patient.PatientBaseDto;
 import com.balybus.galaxy.patient.recruit.dto.RecruitRequestDto;
 import com.balybus.galaxy.patient.recruit.dto.RecruitResponseDto;
 import lombok.RequiredArgsConstructor;
